Added initial code/comments for the 'Making Sense of the Data' section
This section requires literally no Svelte so it was just me figuring out how to draw conclusions from the data. I have listed out the stuff we require from the data and then the steps to take to get there. Started by implementing a function that will need to be cleaned up and renamed that just figures out the accumulated weighting of each Player object by their number of appearances and, given multiple arrays of players (ie. a Region), it will return the array that has the highest weighting, it will return multiple arrays if joint weighting.